refactor redirection notification modal
